Refactor into cpp
