+++ bodo/pandas/physical/gpu_read_iceberg.h
@@ -1,13 +1,18 @@
 #pragma once
 
 #include <Python.h>
-#include <arrow/util/key_value_metadata.h>
+#include <arrow/filesystem/filesystem.h>
+#include <mpi.h>
 #include <cudf/io/datasource.hpp>
+#include <cudf/io/parquet.hpp>
+#include <cudf/table/table.hpp>
+#include <cudf/table/table_view.hpp>
 #include <cudf/types.hpp>
 #include <memory>
-#include <set>
-#include <unordered_map>
+#include <string>
 #include <utility>
+#include <vector>
+
 #include "../../libs/_utils.h"
 #include "../../libs/gpu_utils.h"
 #include "../_util.h"
@@ -15,29 +20,6 @@
 #include "duckdb/planner/table_filter.hpp"
 #include "gpu_expression.h"
 #include "operator.h"
-
-#include <longobject.h>
-#include <mpi.h>
-
-#include <cudf/column/column_factories.hpp>
-#include <cudf/concatenate.hpp>
-#include <cudf/copying.hpp>
-#include <cudf/io/parquet.hpp>
-#include <cudf/table/table.hpp>
-#include <cudf/table/table_view.hpp>
-
-#include <arrow/filesystem/filesystem.h>
-#include <arrow/io/file.h>
-#include <arrow/python/api.h>
-#include <parquet/arrow/reader.h>
-#include <parquet/exception.h>
-#include "../io/arrow_compat.h"
-#include "../io/iceberg_helpers.h"
-#include "physical/gpu_read_parquet.h"
-
-#include <parquet/metadata.h>
-#include <string>
-#include <vector>
 
 /**
  * @brief Batch generator for Iceberg on GPU.
@@ -51,1005 +33,56 @@ class GPUIcebergRankBatchGenerator {
         const int64_t snapshot_id, const std::vector<int>& selected_columns,
         duckdb::unique_ptr<duckdb::TableFilterSet> filter_exprs,
         std::size_t target_rows, MPI_Comm comm,
-        std::shared_ptr<arrow::Schema> output_arrow_schema)
-        : target_rows_(target_rows),
-          filter_exprs_(std::move(filter_exprs)),
-          py_pieces(nullptr),
-          py_schema_groups(nullptr),
-          pyarrow_schema(nullptr),
-          py_filesystem(nullptr),
-          selected_columns_(selected_columns),
-          output_arrow_schema(std::move(output_arrow_schema)) {
-        // Get file list on all ranks and only assign parts to GPU-pinned ranks
-
-        MPI_Comm_rank(comm, &rank_);
-        MPI_Comm_size(comm, &size_);
-
-        get_dataset(catalog, table_id, arrow_schema, iceberg_filter,
-                    iceberg_schema, snapshot_id, target_rows, comm);
-
-        if (!is_gpu_rank()) {
-            return;
-        }
-
-        // Distribute pieces
-        distribute_pieces();
-
-        // Initialize scanners (schemas and filters)
-        init_scanners(comm);
-    }
+        std::shared_ptr<arrow::Schema> output_arrow_schema);
 
     std::pair<std::unique_ptr<cudf::table>, bool> next(
-        std::shared_ptr<StreamAndEvent> se) {
-        if (!is_gpu_rank() ||
-            (curr_piece_idx >= pieces_.size() && !leftover_tbl &&
-             (!curr_reader || !curr_reader->has_next()))) {
-            return {empty_table_from_arrow_schema(output_arrow_schema), true};
-        }
+        std::shared_ptr<StreamAndEvent> se);
 
-        std::vector<std::unique_ptr<cudf::table>> gpu_tables;
-        size_t rows_accum = 0;
-
-        // Handle leftover from previous read
-        if (leftover_tbl) {
-            std::size_t n = leftover_tbl->num_rows();
-            if (n <= target_rows_) {
-                rows_accum += n;
-                gpu_tables.emplace_back(std::move(leftover_tbl));
-                leftover_tbl = nullptr;
-            } else {
-                cudf::table_view tv = leftover_tbl->view();
-                auto sliced =
-                    cudf::slice(tv, {0, (int)target_rows_, (int)n}, se->stream);
-                gpu_tables.push_back(std::make_unique<cudf::table>(sliced[0]));
-                leftover_tbl = std::make_unique<cudf::table>(sliced[1]);
-                rows_accum = target_rows_;
-            }
-        }
-
-        while (rows_accum < target_rows_ && curr_piece_idx < pieces_.size()) {
-            if (!curr_reader) {
-                init_next_reader(se->stream);
-            }
-
-            while (curr_reader && curr_reader->has_next() &&
-                   rows_accum < target_rows_) {
-                auto table_and_metadata = curr_reader->read_chunk();
-                auto tbl = std::move(table_and_metadata.tbl);
-
-                if (!tbl || tbl->num_rows() == 0) {
-                    continue;
-                }
-
-                // Handle rows_to_skip (manual slicing)
-                if (rows_to_skip > 0) {
-                    size_t n = tbl->num_rows();
-                    size_t to_skip = std::min((size_t)rows_to_skip, n);
-                    if (to_skip == n) {
-                        rows_to_skip -= (int64_t)n;
-                        continue;
-                    }
-                    std::vector<cudf::size_type> splits = {
-                        (cudf::size_type)to_skip, (cudf::size_type)n};
-                    auto sliced = cudf::slice(tbl->view(), splits, se->stream);
-                    // sliced[1] is the part to keep
-                    tbl = std::make_unique<cudf::table>(sliced[1]);
-                    rows_to_skip -= (int64_t)to_skip;
-                }
-
-                // Evolve table to match output schema
-                auto evolved_tbl = evolve_table(std::move(tbl), se->stream);
-
-                std::size_t tbl_n = evolved_tbl->num_rows();
-                if (rows_accum + tbl_n <= target_rows_) {
-                    rows_accum += tbl_n;
-                    gpu_tables.push_back(std::move(evolved_tbl));
-                } else {
-                    std::size_t need = target_rows_ - rows_accum;
-                    cudf::table_view tv = evolved_tbl->view();
-                    auto sliced =
-                        cudf::slice(tv, {0, (int)need, (int)tbl_n}, se->stream);
-                    gpu_tables.push_back(
-                        std::make_unique<cudf::table>(sliced[0]));
-                    leftover_tbl = std::make_unique<cudf::table>(sliced[1]);
-                    rows_accum = target_rows_;
-                }
-            }
-
-            if (curr_reader && !curr_reader->has_next()) {
-                curr_reader.reset();
-                // Move to next set of pieces (already handled by
-                // init_next_reader)
-            }
-        }
-
-        if (gpu_tables.empty()) {
-            return {empty_table_from_arrow_schema(output_arrow_schema), true};
-        }
-
-        std::vector<cudf::table_view> views;
-        views.reserve(gpu_tables.size());
-        for (auto& tptr : gpu_tables) {
-            views.emplace_back(tptr->view());
-        }
-
-        std::unique_ptr<cudf::table> batch;
-        if (views.size() == 1) {
-            batch = std::move(gpu_tables[0]);
-        } else {
-            batch = cudf::concatenate(views, se->stream);
-        }
-
-        bool finished = (curr_piece_idx >= pieces_.size() && !leftover_tbl &&
-                         (!curr_reader || !curr_reader->has_next()));
-
-        return {std::move(batch), finished};
-    }
-
-    void ReportMetrics(std::vector<MetricBase>& metrics_out) {
-        MetricBase::StatValue pieces = pieces_.size();
-        metrics_out.emplace_back(StatMetric("num_pieces", pieces));
-        metrics_out.emplace_back(
-            TimerMetric("evolve_time", this->evolve_time_));
-    }
+    void ReportMetrics(std::vector<MetricBase>& metrics_out);
 
    private:
-    void init_next_reader(rmm::cuda_stream_view stream) {
-        if (curr_piece_idx >= pieces_.size()) {
-            return;
-        }
+    void init_next_reader(rmm::cuda_stream_view stream);
 
-        // Process one piece at a time to avoid schema mismatch across files
-        // within the same schema group.
-        int64_t schema_group_idx = pieces_[curr_piece_idx].schema_group_idx;
-        std::string path = pieces_[curr_piece_idx].path;
-        curr_piece_idx++;
-
-        // Get schema group info
-        if (schema_group_idx < 0 ||
-            (size_t)schema_group_idx >= scanner_read_schemas.size()) {
-            throw std::runtime_error(
-                "GPUIcebergRankBatchGenerator::init_next_reader: schema group "
-                "index out of bounds");
-        }
-        curr_read_schema = scanner_read_schemas[schema_group_idx];
-
-        // Identify all columns to read: those in selected_columns_ PLUS those
-        // needed for filters.
-        std::set<int> columns_to_read_set(selected_columns_.begin(),
-                                          selected_columns_.end());
-        for (const auto& pair : filter_exprs_->filters) {
-            columns_to_read_set.insert(pair.first);
-        }
-
-        // Identify existing columns and their file names
-        std::vector<std::string> file_col_names;
-        curr_col_mapping.clear();
-        std::map<int, int> col_to_file_idx;
-
-        for (int col_idx : columns_to_read_set) {
-            std::string name = curr_read_schema->field(col_idx)->name();
-            if (!name.starts_with("_BODO_TEMP_")) {
-                col_to_file_idx[col_idx] = (int)file_col_names.size();
-                file_col_names.push_back(name);
-            }
-        }
-
-        // Map selected columns to their indices in the read table
-        for (size_t i = 0; i < selected_columns_.size(); i++) {
-            int col_idx = selected_columns_[i];
-            if (col_to_file_idx.count(col_idx)) {
-                curr_col_mapping.emplace_back(col_to_file_idx[col_idx], i);
-            } else {
-                // Column doesn't exist in file
-                curr_col_mapping.emplace_back(-1, i);
-            }
-        }
-
-        // Rebuild filter AST only when the schema group changes.
-        if (last_filter_schema_group_idx != schema_group_idx) {
-            filter_ast_tree = cudf::ast::tree();
-            filter_scalars.clear();
-
-            // Check if any filter column was excluded from reading (is a
-            // _BODO_TEMP_ placeholder for a field missing from this file).
-            // All values for such columns are NULL, so the filter is always
-            // FALSE.
-            bool filter_col_missing = false;
-            for (const auto& pair : filter_exprs_->filters) {
-                if (!col_to_file_idx.count(pair.first)) {
-                    filter_col_missing = true;
-                    break;
-                }
-            }
-
-            if (filter_col_missing) {
-                auto false_scalar =
-                    std::make_unique<cudf::numeric_scalar<bool>>(false, true);
-                filter_scalars.push_back(std::move(false_scalar));
-                filter_ast_tree.push(cudf::ast::literal(
-                    *static_cast<cudf::numeric_scalar<bool>*>(
-                        filter_scalars.back().get())));
-                filter_ast_tree.push(cudf::ast::operation(
-                    cudf::ast::ast_operator::IDENTITY, filter_ast_tree.back()));
-            } else {
-                // 1. DuckDB filters -> cuDF AST
-                auto filter_exprs_copy = filter_exprs_->Copy();
-                tableFilterSetToCudfAST(*filter_exprs_copy,
-                                        curr_read_schema->field_names(),
-                                        filter_ast_tree, filter_scalars);
-                const cudf::ast::expression* duckdb_filter_root =
-                    filter_ast_tree.size() > 0 ? &filter_ast_tree.back()
-                                               : nullptr;
-
-                // 2. Iceberg filter -> cuDF AST
-                // Build field_id -> col_idx map
-                std::map<int, int> field_id_to_col_idx;
-                PyObject* field_ids_tuple =
-                    this->scanner_field_ids[schema_group_idx].get();
-                for (int i = 0; i < (int)PyTuple_Size(field_ids_tuple); i++) {
-                    PyObject* item = PyTuple_GetItem(field_ids_tuple, i);
-                    int field_id = -1;
-                    if (PyLong_Check(item)) {
-                        field_id = (int)PyLong_AsLong(item);
-                    } else if (PyTuple_Check(item)) {
-                        field_id = (int)PyLong_AsLong(PyTuple_GetItem(item, 0));
-                    }
-                    if (field_id != -1) {
-                        field_id_to_col_idx[field_id] = i;
-                    }
-                }
-
-                build_pyiceberg_cudf_ast_node(
-                    this->iceberg_filter_cudf_ast.get(), field_id_to_col_idx,
-                    curr_read_schema, filter_ast_tree, filter_scalars, stream);
-                const cudf::ast::expression* iceberg_filter_root =
-                    &filter_ast_tree.back();
-
-                // 3. AND them together
-                if (duckdb_filter_root) {
-                    filter_ast_tree.push(cudf::ast::operation(
-                        cudf::ast::ast_operator::LOGICAL_AND,
-                        *duckdb_filter_root, *iceberg_filter_root));
-                } else {
-                    // No duckdb filter, just use iceberg filter
-                }
-            }
-            last_filter_schema_group_idx = schema_group_idx;
-        }
-
-        cudf::io::parquet_reader_options opts =
-            cudf::io::parquet_reader_options::builder(cudf::io::source_info());
-        opts.set_column_names(file_col_names);
-
-        if (filter_ast_tree.size() > 0) {
-            opts.set_filter(filter_ast_tree.back());
-        }
-
-        std::vector<std::unique_ptr<cudf::io::datasource>> sources;
-        if (filesystem_->type_name() != "local") {
-            std::shared_ptr<arrow::io::RandomAccessFile> arrow_file =
-                filesystem_->OpenInputFile(path).ValueOrDie();
-            sources.push_back(
-                std::make_unique<arrow_file_datasource>(arrow_file));
-        } else {
-            sources.push_back(cudf::io::datasource::create(path));
-        }
-
-        curr_reader = std::make_unique<cudf::io::chunked_parquet_reader>(
-            0,  // chunk_read_limit, 0 for default
-            std::move(sources), std::vector<cudf::io::parquet::FileMetaData>(),
-            opts, stream);
-    }
-
-    // Helper: rebuild a column from released contents after child evolution.
-    // Shares the null-mask + column-construction boilerplate between list
-    // and map evolution paths.
     static std::unique_ptr<cudf::column> make_column_from_contents(
         cudf::column::contents&& col_data, cudf::size_type num_row,
         cudf::size_type null_count, cudf::data_type type,
         std::vector<std::unique_ptr<cudf::column>>&& children,
-        rmm::cuda_stream_view stream) {
-        rmm::device_buffer null_mask = col_data.null_mask
-                                           ? std::move(*col_data.null_mask)
-                                           : rmm::device_buffer{0, stream};
-        return std::make_unique<cudf::column>(
-            type, num_row, std::move(*col_data.data), std::move(null_mask),
-            null_count, std::move(children));
-    }
+        rmm::cuda_stream_view stream);
 
     std::unique_ptr<cudf::column> evolve_column(
         std::unique_ptr<cudf::column> col,
         const std::shared_ptr<arrow::Field>& source_field,
         const std::shared_ptr<arrow::Field>& target_field,
-        rmm::cuda_stream_view stream) {
-        // Verify that the Iceberg field ID is the same.
-        int source_field_iceberg_field_id = get_iceberg_field_id(source_field);
-        int target_field_iceberg_field_id = get_iceberg_field_id(target_field);
-        if (source_field_iceberg_field_id != target_field_iceberg_field_id) {
-            throw std::runtime_error(
-                "GPUIcebergRankBatchGenerator::evolve_column: Iceberg field ID "
-                "of the source (" +
-                std::to_string(source_field_iceberg_field_id) +
-                ") and target (" +
-                std::to_string(target_field_iceberg_field_id) +
-                ") fields do not match!");
-        }
-
-        if (arrow::is_list(source_field->type()->id())) {
-            // Recurse on the value field
-            const std::shared_ptr<arrow::Field>& source_value_field =
-                std::dynamic_pointer_cast<arrow::BaseListType>(
-                    source_field->type())
-                    ->value_field();
-            const std::shared_ptr<arrow::Field>& target_value_field =
-                std::dynamic_pointer_cast<arrow::BaseListType>(
-                    target_field->type())
-                    ->value_field();
-
-            cudf::size_type num_row = col->size();
-            cudf::size_type null_count = col->null_count();
-            cudf::data_type type = col->type();
-            cudf::column::contents col_data = col->release();
-
-            std::unique_ptr<cudf::column> evolved_values_column =
-                evolve_column(std::move(col_data.children[0]),
-                              source_value_field, target_value_field, stream);
-
-            std::vector<std::unique_ptr<cudf::column>> children;
-            children.push_back(std::move(evolved_values_column));
-
-            return make_column_from_contents(std::move(col_data), num_row,
-                                             null_count, type,
-                                             std::move(children), stream);
-
-        } else if (source_field->type()->id() == arrow::Type::MAP) {
-            // Recurse on the key and item fields
-            const std::shared_ptr<arrow::Field>& source_key_field =
-                std::dynamic_pointer_cast<arrow::MapType>(source_field->type())
-                    ->key_field();
-            const std::shared_ptr<arrow::Field>& target_key_field =
-                std::dynamic_pointer_cast<arrow::MapType>(target_field->type())
-                    ->key_field();
-            const std::shared_ptr<arrow::Field>& source_item_field =
-                std::dynamic_pointer_cast<arrow::MapType>(source_field->type())
-                    ->item_field();
-            const std::shared_ptr<arrow::Field>& target_item_field =
-                std::dynamic_pointer_cast<arrow::MapType>(target_field->type())
-                    ->item_field();
-
-            cudf::size_type num_row = col->size();
-            cudf::size_type null_count = col->null_count();
-            cudf::data_type type = col->type();
-            cudf::column::contents col_data = col->release();
-
-            // col_data.children[0] is a Struct column containing keys and items
-            std::unique_ptr<cudf::column> struct_col =
-                std::move(col_data.children[0]);
-            cudf::size_type struct_num_row = struct_col->size();
-            cudf::size_type struct_null_count = struct_col->null_count();
-            cudf::column::contents struct_data = struct_col->release();
-
-            std::unique_ptr<cudf::column> evolved_keys_column =
-                evolve_column(std::move(struct_data.children[0]),
-                              source_key_field, target_key_field, stream);
-            std::unique_ptr<cudf::column> evolved_items_column =
-                evolve_column(std::move(struct_data.children[1]),
-                              source_item_field, target_item_field, stream);
-
-            std::vector<std::unique_ptr<cudf::column>> evolved_struct_children;
-            evolved_struct_children.push_back(std::move(evolved_keys_column));
-            evolved_struct_children.push_back(std::move(evolved_items_column));
-
-            rmm::device_buffer struct_null_mask =
-                struct_data.null_mask ? std::move(*struct_data.null_mask)
-                                      : rmm::device_buffer{0, stream};
-
-            std::unique_ptr<cudf::column> evolved_struct_col =
-                cudf::make_structs_column(
-                    struct_num_row, std::move(evolved_struct_children),
-                    struct_null_count, std::move(struct_null_mask), stream);
-
-            std::vector<std::unique_ptr<cudf::column>> children;
-            children.push_back(std::move(evolved_struct_col));
-
-            return make_column_from_contents(std::move(col_data), num_row,
-                                             null_count, type,
-                                             std::move(children), stream);
-
-        } else if (target_field->type()->id() == arrow::Type::STRUCT) {
-            // Recursive evolution for Struct
-            auto source_type = std::dynamic_pointer_cast<arrow::StructType>(
-                source_field->type());
-            auto target_type = std::dynamic_pointer_cast<arrow::StructType>(
-                target_field->type());
-
-            cudf::size_type num_row = col->size();
-            cudf::size_type null_count = col->null_count();
-            cudf::column::contents col_data = col->release();
-
-            std::vector<std::unique_ptr<cudf::column>> source_children =
-                std::move(col_data.children);
-            std::vector<std::unique_ptr<cudf::column>> evolved_children;
-
-            std::unordered_map<int, int> source_id_to_idx;
-            for (int i = 0; i < source_type->num_fields(); i++) {
-                source_id_to_idx[get_iceberg_field_id(source_type->field(i))] =
-                    i;
-            }
-
-            for (int i = 0; i < target_type->num_fields(); i++) {
-                auto target_child_field = target_type->field(i);
-                int id = get_iceberg_field_id(target_child_field);
-                if (source_id_to_idx.contains(id)) {
-                    int source_idx = source_id_to_idx[id];
-                    evolved_children.push_back(
-                        evolve_column(std::move(source_children[source_idx]),
-                                      source_type->field(source_idx),
-                                      target_child_field, stream));
-                } else {
-                    // Add null child
-                    auto null_scalar = arrow_scalar_to_cudf(
-                        arrow::MakeNullScalar(target_child_field->type()),
-                        stream);
-                    evolved_children.push_back(cudf::make_column_from_scalar(
-                        *null_scalar, num_row, stream));
-                }
-            }
-
-            rmm::device_buffer null_mask = col_data.null_mask
-                                               ? std::move(*col_data.null_mask)
-                                               : rmm::device_buffer{0, stream};
-
-            return cudf::make_structs_column(
-                num_row, std::move(evolved_children), null_count,
-                std::move(null_mask), stream);
-        }
-
-        // For primitive types, cast to target type if needed
-        auto target_cudf_type = arrow_to_cudf_type(target_field->type());
-        if (col->type() != target_cudf_type) {
-            return cudf::cast(col->view(), target_cudf_type, stream);
-        }
-        return col;
-    }
+        rmm::cuda_stream_view stream);
 
     std::unique_ptr<cudf::table> evolve_table(std::unique_ptr<cudf::table> tbl,
-                                              rmm::cuda_stream_view stream) {
-        time_pt start_evolve = start_timer();
-        // We need the read_schema to get the source fields for evolution
-        // We can store it in curr_read_schema
-        std::vector<std::unique_ptr<cudf::column>> output_cols;
-        output_cols.reserve(selected_columns_.size());
+                                              rmm::cuda_stream_view stream);
 
-        size_t num_rows = tbl->num_rows();
-        auto released_cols = tbl->release();
-
-        for (size_t i = 0; i < selected_columns_.size(); i++) {
-            int col_idx = selected_columns_[i];
-            const auto& mapping = curr_col_mapping[i];
-            auto target_field = output_arrow_schema->field(i);
-
-            if (mapping.first == -1) {
-                // Add null column
-                auto arrow_type = target_field->type();
-                auto null_scalar = arrow_scalar_to_cudf(
-                    arrow::MakeNullScalar(arrow_type), stream);
-                output_cols.push_back(cudf::make_column_from_scalar(
-                    *null_scalar, num_rows, stream));
-            } else {
-                auto source_field = curr_read_schema->field(col_idx);
-                output_cols.push_back(
-                    evolve_column(std::move(released_cols[mapping.first]),
-                                  source_field, target_field, stream));
-            }
-        }
-
-        auto ret = std::make_unique<cudf::table>(std::move(output_cols));
-        this->evolve_time_ += end_timer(start_evolve);
-        return ret;
-    }
-
-    // Helper: push an identity(boolean) node onto the filter AST.
     static void push_cudf_identity(
         bool value, cudf::ast::tree& filter_ast_tree,
-        std::vector<std::unique_ptr<cudf::scalar>>& filter_scalars) {
-        auto literal_value =
-            std::make_unique<cudf::numeric_scalar<bool>>(value, true);
-        filter_scalars.push_back(std::move(literal_value));
-        filter_ast_tree.push(
-            cudf::ast::literal(*static_cast<cudf::numeric_scalar<bool>*>(
-                filter_scalars.back().get())));
-        filter_ast_tree.push(cudf::ast::operation(
-            cudf::ast::ast_operator::IDENTITY, filter_ast_tree.back()));
-    }
+        std::vector<std::unique_ptr<cudf::scalar>>& filter_scalars);
 
-    // Helper: convert a Python literal value and push it as a cuDF AST literal
-    // node. This follows the same pattern as duckdbValuetoCudfLiteral:
-    // create a typed scalar, store it in filter_scalars, then push a
-    // cudf::ast::literal via a static_cast back to the concrete type.
     static void push_literal_to_cudf_ast(
         PyObject* value_py, const std::shared_ptr<arrow::DataType>& type,
         cudf::ast::tree& filter_ast_tree,
-        std::vector<std::unique_ptr<cudf::scalar>>& filter_scalars) {
-        bool is_null = (value_py == Py_None);
-        switch (type->id()) {
-            case arrow::Type::BOOL: {
-                auto literal_value =
-                    std::make_unique<cudf::numeric_scalar<bool>>(
-                        is_null ? false : (value_py == Py_True), !is_null);
-                filter_scalars.push_back(std::move(literal_value));
-                filter_ast_tree.push(cudf::ast::literal(
-                    *static_cast<cudf::numeric_scalar<bool>*>(
-                        filter_scalars.back().get())));
-                return;
-            }
-            case arrow::Type::INT8: {
-                auto literal_value =
-                    std::make_unique<cudf::numeric_scalar<int8_t>>(
-                        is_null ? 0 : (int8_t)PyLong_AsLong(value_py),
-                        !is_null);
-                filter_scalars.push_back(std::move(literal_value));
-                filter_ast_tree.push(cudf::ast::literal(
-                    *static_cast<cudf::numeric_scalar<int8_t>*>(
-                        filter_scalars.back().get())));
-                return;
-            }
-            case arrow::Type::INT16: {
-                auto literal_value =
-                    std::make_unique<cudf::numeric_scalar<int16_t>>(
-                        is_null ? 0 : (int16_t)PyLong_AsLong(value_py),
-                        !is_null);
-                filter_scalars.push_back(std::move(literal_value));
-                filter_ast_tree.push(cudf::ast::literal(
-                    *static_cast<cudf::numeric_scalar<int16_t>*>(
-                        filter_scalars.back().get())));
-                return;
-            }
-            case arrow::Type::INT32: {
-                auto literal_value =
-                    std::make_unique<cudf::numeric_scalar<int32_t>>(
-                        is_null ? 0 : (int32_t)PyLong_AsLong(value_py),
-                        !is_null);
-                filter_scalars.push_back(std::move(literal_value));
-                filter_ast_tree.push(cudf::ast::literal(
-                    *static_cast<cudf::numeric_scalar<int32_t>*>(
-                        filter_scalars.back().get())));
-                return;
-            }
-            case arrow::Type::INT64: {
-                auto literal_value =
-                    std::make_unique<cudf::numeric_scalar<int64_t>>(
-                        is_null ? 0 : (int64_t)PyLong_AsLongLong(value_py),
-                        !is_null);
-                filter_scalars.push_back(std::move(literal_value));
-                filter_ast_tree.push(cudf::ast::literal(
-                    *static_cast<cudf::numeric_scalar<int64_t>*>(
-                        filter_scalars.back().get())));
-                return;
-            }
-            case arrow::Type::FLOAT: {
-                auto literal_value =
-                    std::make_unique<cudf::numeric_scalar<float>>(
-                        is_null ? 0.0f : (float)PyFloat_AsDouble(value_py),
-                        !is_null);
-                filter_scalars.push_back(std::move(literal_value));
-                filter_ast_tree.push(cudf::ast::literal(
-                    *static_cast<cudf::numeric_scalar<float>*>(
-                        filter_scalars.back().get())));
-                return;
-            }
-            case arrow::Type::DOUBLE: {
-                auto literal_value =
-                    std::make_unique<cudf::numeric_scalar<double>>(
-                        is_null ? 0.0 : PyFloat_AsDouble(value_py), !is_null);
-                filter_scalars.push_back(std::move(literal_value));
-                filter_ast_tree.push(cudf::ast::literal(
-                    *static_cast<cudf::numeric_scalar<double>*>(
-                        filter_scalars.back().get())));
-                return;
-            }
-            case arrow::Type::STRING: {
-                const char* str = is_null ? "" : PyUnicode_AsUTF8(value_py);
-                auto literal_value = std::make_unique<cudf::string_scalar>(
-                    std::string(str), !is_null);
-                filter_scalars.push_back(std::move(literal_value));
-                filter_ast_tree.push(
-                    cudf::ast::literal(*static_cast<cudf::string_scalar*>(
-                        filter_scalars.back().get())));
-                return;
-            }
-            default:
-                throw std::runtime_error(
-                    "push_literal_to_cudf_ast: unsupported type: " +
-                    type->ToString());
-        }
-    }
+        std::vector<std::unique_ptr<cudf::scalar>>& filter_scalars);
 
-    // Recursively walk a pyiceberg cuDF AST filter tree (nested Python tuples)
-    // and push cuDF AST nodes onto filter_ast_tree. The root node of this
-    // subtree ends up at filter_ast_tree.back().
     void build_pyiceberg_cudf_ast_node(
         PyObject* node, const std::map<int, int>& field_id_to_col_idx,
         const std::shared_ptr<arrow::Schema>& read_schema,
         cudf::ast::tree& filter_ast_tree,
         std::vector<std::unique_ptr<cudf::scalar>>& filter_scalars,
-        rmm::cuda_stream_view stream) {
-        if (!PyTuple_Check(node)) {
-            push_cudf_identity(true, filter_ast_tree, filter_scalars);
-            return;
-        }
-
-        PyObject* op_py = PyTuple_GetItem(node, 0);
-        if (!op_py || !PyUnicode_Check(op_py)) {
-            push_cudf_identity(true, filter_ast_tree, filter_scalars);
-            return;
-        }
-        const char* op = PyUnicode_AsUTF8(op_py);
-
-        // Always-true / always-false leaf nodes
-        if (strcmp(op, "true") == 0) {
-            push_cudf_identity(true, filter_ast_tree, filter_scalars);
-            return;
-        }
-
-        if (strcmp(op, "false") == 0) {
-            push_cudf_identity(false, filter_ast_tree, filter_scalars);
-            return;
-        }
-
-        // Resolve field_id -> column index and name
-        auto get_col_info =
-            [&](PyObject* node_inner) -> std::pair<int, std::string> {
-            PyObject* field_id_py = PyTuple_GetItem(node_inner, 1);
-            int field_id = (int)PyLong_AsLong(field_id_py);
-            auto it = field_id_to_col_idx.find(field_id);
-            if (it == field_id_to_col_idx.end()) {
-                throw std::runtime_error(
-                    "build_pyiceberg_cudf_ast_node: field ID " +
-                    std::to_string(field_id) + " not found in schema group");
-            }
-            int col_idx = it->second;
-            return {col_idx, read_schema->field(col_idx)->name()};
-        };
-
-        // is_null / is_not_null
-        if (strcmp(op, "is_null") == 0 || strcmp(op, "is_not_null") == 0) {
-            auto [col_idx, col_name] = get_col_info(node);
-            cudf::ast::column_name_reference col_ref(col_name);
-            filter_ast_tree.push(col_ref);
-            cudf::ast::operation expr = cudf::ast::operation(
-                cudf::ast::ast_operator::IS_NULL, filter_ast_tree.back());
-            filter_ast_tree.push(expr);
-            if (strcmp(op, "is_not_null") == 0) {
-                filter_ast_tree.push(cudf::ast::operation(
-                    cudf::ast::ast_operator::NOT, filter_ast_tree.back()));
-            }
-            return;
-        }
-
-        // Comparison ops: eq, neq, gt, gte, lt, lte
-        static const std::unordered_map<std::string, cudf::ast::ast_operator>
-            cmp_ops = {
-                {"eq", cudf::ast::ast_operator::EQUAL},
-                {"neq", cudf::ast::ast_operator::NOT_EQUAL},
-                {"gt", cudf::ast::ast_operator::GREATER},
-                {"gte", cudf::ast::ast_operator::GREATER_EQUAL},
-                {"lt", cudf::ast::ast_operator::LESS},
-                {"lte", cudf::ast::ast_operator::LESS_EQUAL},
-            };
-        auto cmp_it = cmp_ops.find(op);
-        if (cmp_it != cmp_ops.end()) {
-            auto [col_idx, col_name] = get_col_info(node);
-            PyObject* value_py = PyTuple_GetItem(node, 2);
-            auto col_type = read_schema->field(col_idx)->type();
-
-            cudf::ast::column_name_reference col_ref(col_name);
-            filter_ast_tree.push(col_ref);
-            push_literal_to_cudf_ast(value_py, col_type, filter_ast_tree,
-                                     filter_scalars);
-
-            cudf::ast::operation expr = cudf::ast::operation(
-                cmp_it->second, filter_ast_tree[filter_ast_tree.size() - 2],
-                filter_ast_tree.back());
-            filter_ast_tree.push(expr);
-            return;
-        }
-
-        // IN / NOT IN
-        if (strcmp(op, "in") == 0 || strcmp(op, "not_in") == 0) {
-            auto [col_idx, col_name] = get_col_info(node);
-            auto col_type = read_schema->field(col_idx)->type();
-            PyObject* values_list = PyTuple_GetItem(node, 2);
-            Py_ssize_t n_values = PyList_Size(values_list);
-
-            if (n_values == 0) {
-                // Empty IN -> no rows match. Push IDENTITY(false).
-                push_cudf_identity(false, filter_ast_tree, filter_scalars);
-                return;
-            }
-
-            // Expand to OR of EQUAL comparisons
-            PyObject* first_val = PyList_GetItem(values_list, 0);
-            cudf::ast::column_name_reference col_ref(col_name);
-            filter_ast_tree.push(col_ref);
-            push_literal_to_cudf_ast(first_val, col_type, filter_ast_tree,
-                                     filter_scalars);
-
-            const cudf::ast::expression* acc =
-                &filter_ast_tree.push(cudf::ast::operation(
-                    cudf::ast::ast_operator::EQUAL,
-                    filter_ast_tree[filter_ast_tree.size() - 3],
-                    filter_ast_tree.back()));
-
-            for (Py_ssize_t i = 1; i < n_values; i++) {
-                PyObject* val = PyList_GetItem(values_list, i);
-                cudf::ast::column_name_reference col_ref_i(col_name);
-                filter_ast_tree.push(col_ref_i);
-                push_literal_to_cudf_ast(val, col_type, filter_ast_tree,
-                                         filter_scalars);
-
-                const cudf::ast::expression* eq_node =
-                    &filter_ast_tree.push(cudf::ast::operation(
-                        cudf::ast::ast_operator::EQUAL,
-                        filter_ast_tree[filter_ast_tree.size() - 3],
-                        filter_ast_tree.back()));
-
-                acc = &filter_ast_tree.push(cudf::ast::operation(
-                    cudf::ast::ast_operator::LOGICAL_OR, *acc, *eq_node));
-            }
-
-            if (strcmp(op, "not_in") == 0) {
-                filter_ast_tree.push(cudf::ast::operation(
-                    cudf::ast::ast_operator::NOT, filter_ast_tree.back()));
-            }
-            return;
-        }
-
-        // NOT
-        if (strcmp(op, "not") == 0) {
-            PyObject* child = PyTuple_GetItem(node, 1);
-            build_pyiceberg_cudf_ast_node(child, field_id_to_col_idx,
-                                          read_schema, filter_ast_tree,
-                                          filter_scalars, stream);
-            filter_ast_tree.push(cudf::ast::operation(
-                cudf::ast::ast_operator::NOT, filter_ast_tree.back()));
-            return;
-        }
-
-        // AND / OR
-        if (strcmp(op, "and") == 0 || strcmp(op, "or") == 0) {
-            PyObject* left = PyTuple_GetItem(node, 1);
-            PyObject* right = PyTuple_GetItem(node, 2);
-
-            build_pyiceberg_cudf_ast_node(left, field_id_to_col_idx,
-                                          read_schema, filter_ast_tree,
-                                          filter_scalars, stream);
-            const cudf::ast::expression* left_root = &filter_ast_tree.back();
-            build_pyiceberg_cudf_ast_node(right, field_id_to_col_idx,
-                                          read_schema, filter_ast_tree,
-                                          filter_scalars, stream);
-            const cudf::ast::expression* right_root = &filter_ast_tree.back();
-
-            cudf::ast::ast_operator comb_op =
-                (strcmp(op, "and") == 0) ? cudf::ast::ast_operator::LOGICAL_AND
-                                         : cudf::ast::ast_operator::LOGICAL_OR;
-
-            filter_ast_tree.push(
-                cudf::ast::operation(comb_op, *left_root, *right_root));
-            return;
-        }
-
-        // Unknown op, treat as no-op (true)
-        push_cudf_identity(true, filter_ast_tree, filter_scalars);
-    }
+        rmm::cuda_stream_view stream);
 
     void get_dataset(PyObject* catalog, const std::string& table_id,
                      const std::shared_ptr<arrow::Schema>& arrow_schema,
                      PyObject* iceberg_filter, PyObject* iceberg_schema,
                      int64_t snapshot_id, int64_t tot_rows_to_read,
-                     MPI_Comm comm) {
-        // import bodo.io.iceberg
-        PyObjectPtr iceberg_mod = PyImport_ImportModule("bodo.io.iceberg");
-        if (PyErr_Occurred()) {
-            throw std::runtime_error(
-                "GPUIcebergRankBatchGenerator::get_dataset: failed to import "
-                "bodo.io.iceberg module");
-        }
+                     MPI_Comm comm);
 
-        // For now, we don't support dict-encoded strings on GPU yet (or we'll
-        // handle them as strings)
-        PyObjectPtr str_as_dict_cols_py = PyList_New(0);
+    void distribute_pieces();
 
-        // We use piece-level read for GPU to avoid expensive exact row counting
-        // on CPU if possible.
-        PyObject* force_row_level_py = Py_False;
-
-        this->pyarrow_schema = arrow::py::wrap_schema(arrow_schema);
-
-        // We need to & the iceberg_filter with converted duckdb table filters
-        // to apply the filters at the file level.
-        PyObjectPtr duckdb_iceberg_filter =
-            duckdbFilterSetToPyicebergFilter(*filter_exprs_, arrow_schema);
-
-        // Perform the python & to combine the filters
-        PyObjectPtr py_iceberg_filter_and_duckdb_filter =
-            PyObjectPtr(PyObject_CallMethod(duckdb_iceberg_filter.get(),
-                                            "__and__", "O", iceberg_filter));
-        if (!py_iceberg_filter_and_duckdb_filter) {
-            throw std::runtime_error(
-                "failed to combine iceberg filter with duckdb table filters");
-        }
-
-        // We need to convert the combined pyiceberg iceberg filter to an Arrow
-        // filter format string so it can be applied at the row level in
-        // addition to the file level.
-        //  import bodo.io.iceberg.common
-        PyObjectPtr iceberg_common_mod =
-            PyImport_ImportModule("bodo.io.iceberg.common");
-        if (PyErr_Occurred()) {
-            throw std::runtime_error(
-                "failed to import bodo.io.iceberg.common module");
-        }
-
-        // Convert iceberg filter to cuDF AST
-        PyObjectPtr cudf_ast_func = PyObject_GetAttrString(
-            iceberg_common_mod.get(), "pyiceberg_filter_to_cudf_ast");
-        if (!cudf_ast_func || !PyCallable_Check(cudf_ast_func)) {
-            throw std::runtime_error(
-                "failed to get pyiceberg_filter_to_cudf_ast function");
-        }
-        this->iceberg_filter_cudf_ast = PyObjectPtr(
-            PyObject_CallFunctionObjArgs(cudf_ast_func.get(), iceberg_filter,
-                                         iceberg_schema, Py_True, nullptr));
-        if (!this->iceberg_filter_cudf_ast) {
-            throw std::runtime_error(
-                "failed to convert iceberg filter to cuDF AST");
-        }
-
-        PyObjectPtr convert_func = PyObject_GetAttrString(
-            iceberg_common_mod.get(),
-            "pyiceberg_filter_to_pyarrow_format_str_and_scalars");
-        if (!convert_func || !PyCallable_Check(convert_func)) {
-            throw std::runtime_error(
-                "failed to get convert_iceberg_filter_to_arrow function from "
-                "bodo.io.iceberg.common module");
-        }
-
-        // call python function to convert the combined pyiceberg filter
-        PyObjectPtr filter_f_str_and_scalars =
-            PyObjectPtr(PyObject_CallFunctionObjArgs(
-                convert_func.get(), py_iceberg_filter_and_duckdb_filter.get(),
-                iceberg_schema, Py_True, nullptr));
-        if (!filter_f_str_and_scalars) {
-            throw std::runtime_error(
-                "failed to convert pyiceberg filter to arrow filter format "
-                "string");
-        }
-
-        // The result is a tuple of (iceberg_filter_f_str, filter_scalars)
-        PyObject* iceberg_filter_f_str_py =
-            PyTuple_GetItem(filter_f_str_and_scalars.get(), 0);
-        PyObject* filter_scalars_py =
-            PyTuple_GetItem(filter_f_str_and_scalars.get(), 1);
-
-        PyObjectPtr ds = PyObject_CallMethod(
-            iceberg_mod.get(), "get_iceberg_pq_dataset", "OsOOOsOOLL", catalog,
-            table_id.c_str(), this->pyarrow_schema.get(),
-            str_as_dict_cols_py.get(),
-            py_iceberg_filter_and_duckdb_filter.get(),
-            iceberg_filter_f_str_py != Py_None
-                ? PyUnicode_AsUTF8(iceberg_filter_f_str_py)
-                : "",
-            filter_scalars_py, force_row_level_py,
-            static_cast<long long>(snapshot_id),
-            static_cast<long long>(tot_rows_to_read));
-
-        if (ds == nullptr || PyErr_Occurred()) {
-            throw std::runtime_error(
-                "Error during Iceberg read: Failed to get dataset from Python");
-        }
-
-        this->py_filesystem = PyObject_GetAttrString(ds.get(), "filesystem");
-        ensure_pa_wrappers_imported();
-        if (!(this->py_filesystem == nullptr ||
-              this->py_filesystem == Py_None)) {
-            CHECK_ARROW_AND_ASSIGN(
-                arrow::py::unwrap_filesystem(this->py_filesystem.get()),
-                "Error during Iceberg read: Failed to unwrap Arrow filesystem",
-                this->filesystem_);
-        }
-
-        this->py_pieces = PyObject_GetAttrString(ds.get(), "pieces");
-        this->py_schema_groups =
-            PyObject_GetAttrString(ds.get(), "schema_groups");
-    }
-
-    void distribute_pieces() {
-        // import bodo.io.iceberg.read_parquet
-        PyObjectPtr iceberg_mod =
-            PyImport_ImportModule("bodo.io.iceberg.read_parquet");
-        if (PyErr_Occurred()) {
-            throw std::runtime_error(
-                "GPUIcebergRankBatchGenerator::distribute_pieces: failed to "
-                "import bodo.io.iceberg.read_parquet module");
-        }
-
-        PyObjectPtr rank_py = PyLong_FromLong(rank_);
-        PyObjectPtr size_py = PyLong_FromLong(size_);
-
-        // pieces_myrank_py =
-        // bodo.io.iceberg.read_parquet.distribute_pieces(pieces_py, rank_py,
-        // size_py)
-        PyObjectPtr pieces_myrank_py = PyObject_CallMethod(
-            iceberg_mod.get(), "distribute_pieces", "OOO",
-            this->py_pieces.get(), rank_py.get(), size_py.get());
-        if (pieces_myrank_py == nullptr && PyErr_Occurred()) {
-            throw std::runtime_error(
-                "GPUIcebergRankBatchGenerator::distribute_pieces: "
-                "distribute_pieces call failed");
-        }
-
-        PyObject* piece;
-        PyObjectPtr iterator = PyObject_GetIter(pieces_myrank_py.get());
-        if (iterator == nullptr) {
-            throw std::runtime_error(
-                "GPUIcebergRankBatchGenerator::distribute_pieces(): error "
-                "getting pieces iterator");
-        }
-
-        while ((piece = PyIter_Next(iterator))) {
-            // Add piece to our local list
-            // Extract path, num_rows, schema_group_idx
-            PyObjectPtr p = PyObject_GetAttrString(piece, "path");
-            const char* c_path = PyUnicode_AsUTF8(p.get());
-
-            PyObjectPtr num_rows_piece_py =
-                PyObject_GetAttrString(piece, "_bodo_num_rows");
-            int64_t num_rows_piece = PyLong_AsLongLong(num_rows_piece_py.get());
-
-            PyObjectPtr schema_group_idx_py =
-                PyObject_GetAttrString(piece, "schema_group_idx");
-            int64_t schema_group_idx =
-                PyLong_AsLongLong(schema_group_idx_py.get());
-
-            this->pieces_.push_back({c_path, num_rows_piece, schema_group_idx});
-
-            Py_DECREF(piece);
-        }
-    }
-
-    void init_scanners(MPI_Comm comm) {
-        if (pieces_.empty()) {
-            return;
-        }
-
-        // Extract read schemas directly from schema groups
-        size_t n_schema_groups = PyList_Size(this->py_schema_groups.get());
-        this->scanner_read_schemas.reserve(n_schema_groups);
-        for (size_t i = 0; i < n_schema_groups; i++) {
-            PyObject* schema_group =
-                PyList_GetItem(this->py_schema_groups.get(), i);
-            PyObjectPtr read_schema_py = PyObjectPtr(
-                PyObject_GetAttrString(schema_group, "read_schema"));
-            if (!read_schema_py) {
-                throw std::runtime_error(
-                    "GPUIcebergRankBatchGenerator::init_scanners: failed to "
-                    "get read_schema from schema group");
-            }
-            std::shared_ptr<arrow::Schema> read_schema;
-            CHECK_ARROW_AND_ASSIGN(
-                arrow::py::unwrap_schema(read_schema_py.get()),
-                "GPUIcebergRankBatchGenerator::init_scanners: failed to unwrap "
-                "schema",
-                read_schema);
-            this->scanner_read_schemas.push_back(read_schema);
-            this->scanner_field_ids.emplace_back(
-                PyObject_GetAttrString(schema_group, "iceberg_field_ids"));
-        }
-
-        // rows_to_skip is always 0 for GPU (piece-level read, no offset)
-    }
+    void init_scanners(MPI_Comm comm);
 
     struct IcebergPieceInfo {
         std::string path;
@@ -1124,107 +157,17 @@ class PhysicalGPUReadIceberg : public PhysicalGPUSource {
         const int64_t snapshot_id, const std::vector<int>& selected_columns,
         duckdb::TableFilterSet& filter_exprs,
         duckdb::unique_ptr<duckdb::BoundLimitNode>& limit_val,
-        JoinFilterColStats join_filter_col_stats)
-        : catalog(catalog),
-          table_id(table_id),
-          iceberg_filter(iceberg_filter),
-          iceberg_schema(iceberg_schema),
-          snapshot_id(snapshot_id),
-          filter_exprs(filter_exprs.Copy()),
-          arrow_schema(arrow_schema),
-          selected_columns(selected_columns),
-          join_filter_col_stats(std::move(join_filter_col_stats)) {
-        Py_INCREF(this->catalog);
-        Py_INCREF(this->iceberg_filter);
-        Py_INCREF(this->iceberg_schema);
+        JoinFilterColStats join_filter_col_stats);
 
-        // Initialize schemas and metadata
-        this->output_schema = bodo::Schema::FromArrowSchema(arrow_schema)
-                                  ->Project(selected_columns);
-        // Create a new schema with only the selected columns.
-        std::vector<std::shared_ptr<arrow::Field>> fields;
-        fields.reserve(selected_columns.size());
-        for (int i : selected_columns) {
-            fields.push_back(arrow_schema->field(i));
-        }
-        this->output_arrow_schema =
-            arrow::schema(fields, arrow_schema->metadata());
+    virtual ~PhysicalGPUReadIceberg();
 
-        this->comm = get_gpu_mpi_comm(get_gpu_id());
-    }
-
-    virtual ~PhysicalGPUReadIceberg() {
-        Py_XDECREF(this->catalog);
-        Py_XDECREF(this->iceberg_filter);
-        Py_XDECREF(this->iceberg_schema);
-        if (this->comm != MPI_COMM_NULL) {
-            MPI_Comm_free(&this->comm);
-        }
-    }
-
-    void FinalizeSource() override {
-        std::vector<MetricBase> metrics_out;
-        metrics_out.emplace_back(
-            TimerMetric("produce_time", this->metrics.produce_time));
-        if (this->batch_gen) {
-            this->batch_gen->ReportMetrics(metrics_out);
-        }
-
-        QueryProfileCollector::Default().SubmitOperatorName(getOpId(),
-                                                            ToString());
-        QueryProfileCollector::Default().SubmitOperatorStageTime(
-            QueryProfileCollector::MakeOperatorStageID(getOpId(), 0),
-            this->metrics.init_time);
-        QueryProfileCollector::Default().SubmitOperatorStageTime(
-            QueryProfileCollector::MakeOperatorStageID(getOpId(), 1),
-            this->metrics.produce_time);
-        QueryProfileCollector::Default().RegisterOperatorStageMetrics(
-            QueryProfileCollector::MakeOperatorStageID(getOpId(), 1),
-            std::move(metrics_out));
-        QueryProfileCollector::Default().SubmitOperatorStageRowCounts(
-            QueryProfileCollector::MakeOperatorStageID(getOpId(), 1),
-            this->metrics.rows_read);
-    }
+    void FinalizeSource() override;
 
     std::pair<GPU_DATA, OperatorResult> ProduceBatchGPU(
-        std::shared_ptr<StreamAndEvent> se) override {
-        if (!batch_gen) {
-            time_pt start_init = start_timer();
-            init_batch_gen();
-            this->metrics.init_time += end_timer(start_init);
-        }
+        std::shared_ptr<StreamAndEvent> se) override;
 
-        if (!is_gpu_rank()) {
-            return {GPU_DATA(nullptr, output_arrow_schema, se),
-                    OperatorResult::FINISHED};
-        }
-
-        time_pt start_produce = start_timer();
-        auto next_batch_tup = batch_gen->next(se);
-        auto result = next_batch_tup.second ? OperatorResult::FINISHED
-                                            : OperatorResult::HAVE_MORE_OUTPUT;
-
-        if (next_batch_tup.first) {
-            this->metrics.rows_read += next_batch_tup.first->num_rows();
-        }
-
-        GPU_DATA ret(std::move(next_batch_tup.first), output_arrow_schema, se);
-        this->metrics.produce_time += end_timer(start_produce);
-        return {std::move(ret), result};
-    }
-
-    const std::shared_ptr<bodo::Schema> getOutputSchemaInternal() override {
-        return output_schema;
-    }
+    const std::shared_ptr<bodo::Schema> getOutputSchemaInternal() override;
 
    private:
-    void init_batch_gen() {
-        auto batch_size = get_gpu_streaming_batch_size();
-        this->filter_exprs = this->join_filter_col_stats.insert_filters(
-            std::move(this->filter_exprs), this->selected_columns);
-        batch_gen = std::make_shared<GPUIcebergRankBatchGenerator>(
-            catalog, table_id, iceberg_filter, iceberg_schema, arrow_schema,
-            snapshot_id, selected_columns, std::move(filter_exprs), batch_size,
-            comm, output_arrow_schema);
-    }
+    void init_batch_gen();
 };
